[AUTO BATCH PLUGIN] Add plugin.cpp, seprate the plugin related operations from auto_batch.cpp
